+++ test/port/controller/UserController/whoami.test.ts
@@ -1,5 +1,7 @@
+import { AuthAdapter } from 'app/infra/AuthAdapter';
 import assert from 'assert';
-import { app } from 'egg-mock/bootstrap';
+import dayjs from 'dayjs';
+import { app, mock } from 'egg-mock/bootstrap';
 import { TestUtil } from 'test/TestUtil';
 
 describe('test/port/controller/UserController/whoami.test.ts', () => {
@@ -10,7 +12,7 @@ describe('test/port/controller/UserController/whoami.test.ts', () => {
         .get('/-/whoami')
         .set('authorization', authorization)
         .expect(200);
-      assert.equal(res.body.username, name);
+      assert.deepStrictEqual(res.body, { username: name });
     });
 
     it('should unauthorized', async () => {
@@ -24,6 +26,39 @@ describe('test/port/controller/UserController/whoami.test.ts', () => {
         .set('authorization', 'Bearer foo-token')
         .expect(401);
       assert.equal(res.body.error, '[UNAUTHORIZED] Invalid token');
+    });
+
+    it('should return granular token info', async () => {
+      const { name, email } = await TestUtil.createUser({ name: 'banana' });
+      await TestUtil.createPackage({ name: '@cnpm/banana' });
+      mock(AuthAdapter.prototype, 'ensureCurrentUser', async () => {
+        return {
+          name,
+          email,
+        };
+      });
+      let res = await app.httpRequest()
+        .post('/-/npm/v1/tokens/gat')
+        .send({
+          name: 'apple',
+          description: 'lets play',
+          allowedPackages: [ '@cnpm/banana' ],
+          allowedScopes: [ '@banana' ],
+          expires: 30,
+        })
+        .expect(200);
+
+      res = await app.httpRequest()
+        .get('/-/whoami')
+        .set('authorization', `Bearer ${res.body.token}`)
+        .expect(200);
+
+      assert.equal(res.body.username, name);
+      assert.equal(res.body.name, 'apple');
+      assert.equal(res.body.description, 'lets play');
+      assert.deepEqual(res.body.allowedPackages, [ '@cnpm/banana' ]);
+      assert.deepEqual(res.body.allowedScopes, [ '@banana' ]);
+      assert(dayjs(res.body.expires).isBefore(dayjs().add(30, 'days')));
     });
   });
 });